Change the wifi-reset button to a winfi-config button
When touched while the app is running, the wifi will enter config mode without resetting the data. Hitting the exit button on the front page now also closes the connection and restarts.

Also adds a parameter to the wifi manager with device ID, however, this doesn't seem to work. I've created an issue request on the original library.